fix: example template only worked with probablykasper user
The app crashed with: ::error::Could not resolve to a Repository with the name 'hlinero/vidl'.

It is trying to use the person’s account ${{user.USERNAME}} to retrieve the `vidl` repository but such a repo only belongs to probablykasper. So, for everyone else, the tests will not run.

I am specifying the owner of `vidl` instead of dynamically retrieving it.